domain-search/pyproject.toml
[project]
name = "domain-search"
version = "1.0"
description = "CLI tool to search for available domain names across all known TLDs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dnspython>=2.7.0",
    "httpx>=0.28.1",
    "rich>=13.0.0",
]

[project.scripts]
domain-search = "domain_search.cli:main"

[project.urls]
Homepage = "https://github.com/sasha-computer/domain-search"
Repository = "https://github.com/sasha-computer/domain-search"
Issues = "https://github.com/sasha-computer/domain-search/issues"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.13.0",
    "ty>=0.0.15",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "B",
    "UP",
    "SIM",
]
ignore = []

[tool.ty.environment]
python-version = "3.12"

[tool.ty.src]
include = ["domain_search", "tests"]
exclude = [
    ".venv",
    "build",
    "dist",
]

[tool.ty.rules]
all = "error"
possibly-missing-import = "error"
possibly-missing-attribute = "error"
unresolved-attribute = "error"
unresolved-reference = "error"
